refactor: split marketplace sims
