+++ x/ecocredit/marketplace/simulation/msg_sell.go
@@ -1,0 +1,119 @@
+package simulation
+
+import (
+	"fmt"
+	"math/rand"
+
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	"github.com/cosmos/cosmos-sdk/simapp"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+	"github.com/cosmos/cosmos-sdk/x/simulation"
+
+	"github.com/regen-network/regen-ledger/types/math"
+	"github.com/regen-network/regen-ledger/x/ecocredit"
+	basetypes "github.com/regen-network/regen-ledger/x/ecocredit/base/types/v1"
+	types "github.com/regen-network/regen-ledger/x/ecocredit/marketplace/types/v1"
+	"github.com/regen-network/regen-ledger/x/ecocredit/simulation/utils"
+)
+
+const OpWeightMsgSell = "op_weight_msg_sell" //nolint:gosec
+
+const WeightSell = 100
+
+var TypeMsgSell = types.MsgSell{}.Route()
+
+// SimulateMsgSell generates a Marketplace/MsgSell with random values.
+func SimulateMsgSell(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper,
+	qryClient basetypes.QueryServer) simtypes.Operation {
+	return func(
+		r *rand.Rand, baseApp *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
+	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
+		seller, _ := simtypes.RandomAcc(r, accs)
+		sellerAddr := seller.Address.String()
+
+		ctx := sdk.WrapSDKContext(sdkCtx)
+		class, op, err := utils.GetRandomClass(sdkCtx, r, qryClient, TypeMsgSell)
+		if class == nil {
+			return op, nil, err
+		}
+
+		batchRes, err := qryClient.BatchesByIssuer(ctx, &basetypes.QueryBatchesByIssuerRequest{Issuer: seller.Address.String()})
+		if err != nil {
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSell, err.Error()), nil, err
+		}
+
+		batches := batchRes.Batches
+		if len(batches) == 0 {
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSell, "no credit batches"), nil, nil
+		}
+		max := 1
+		if len(batches) > 1 {
+			max = simtypes.RandIntBetween(r, 1, len(batches))
+		}
+
+		sellOrders := make([]*types.MsgSell_Order, max)
+		for i := 0; i < max; i++ {
+			bal, err := qryClient.Balance(ctx, &basetypes.QueryBalanceRequest{Address: sellerAddr, BatchDenom: batches[i].Denom})
+			if err != nil {
+				return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSell, err.Error()), nil, err
+			}
+
+			exp := sdkCtx.BlockTime().AddDate(1, 0, 0)
+			d, err := math.NewNonNegativeDecFromString(bal.Balance.TradableAmount)
+			if err != nil {
+				return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSell, err.Error()), nil, err
+			}
+
+			if d.IsZero() {
+				return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSell, "no balance"), nil, nil
+			}
+
+			balInt, err := d.Int64()
+			if err != nil {
+				return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgSell, err.Error()), nil, nil
+			}
+
+			quantity := int(balInt)
+			if balInt > 1 {
+				quantity = simtypes.RandIntBetween(r, 1, quantity)
+			}
+
+			askPrice := sdk.NewInt64Coin(sdk.DefaultBondDenom, int64(simtypes.RandIntBetween(r, 1, 50)))
+			sellOrders[i] = &types.MsgSell_Order{
+				BatchDenom:        batches[i].Denom,
+				Quantity:          fmt.Sprintf("%d", quantity),
+				AskPrice:          &askPrice,
+				DisableAutoRetire: r.Int63n(101) <= 30, // 30% chance of disable auto-retire
+				Expiration:        &exp,
+			}
+		}
+
+		msg := &types.MsgSell{
+			Seller: seller.Address.String(),
+			Orders: sellOrders,
+		}
+
+		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, sellerAddr, TypeMsgSell)
+		if spendable == nil {
+			return op, nil, err
+		}
+
+		txCtx := simulation.OperationInput{
+			R:               r,
+			App:             baseApp,
+			TxGen:           simapp.MakeTestEncodingConfig().TxConfig,
+			Cdc:             nil,
+			Msg:             msg,
+			MsgType:         msg.Type(),
+			Context:         sdkCtx,
+			SimAccount:      *account,
+			AccountKeeper:   ak,
+			Bankkeeper:      bk,
+			ModuleName:      ecocredit.ModuleName,
+			CoinsSpentInMsg: spendable,
+		}
+
+		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
+	}
+}
